Remove video player screen
