+++ screens/HomeScreen.js
@@ -17,6 +17,7 @@ import ErrorView from '../components/ErrorView';
 import Colors from '../constants/Colors';
 import Screens from '../constants/Screens';
 import { getIconName } from '../utils/Icons';
+import VideoPlayer from '../components/VideoPlayer';
 
 const HomeScreen = observer(() => {
 	const { rootStore } = useStores();
@@ -59,12 +60,6 @@ const HomeScreen = observer(() => {
 	useEffect(() => {
 		setIsLoading(true);
 	}, [rootStore.settingStore.activeServer]);
-
-	useEffect(() => {
-		if (rootStore.mediaStore.isActive && rootStore.mediaStore.type === 'Video') {
-			navigation.navigate(Screens.VideoPlayerScreen);
-		}
-	}, [rootStore.mediaStore.isActive]);
 
 	useEffect(() => {
 		if (httpErrorStatus) {
@@ -121,58 +116,61 @@ const HomeScreen = observer(() => {
 				}} />
 			)}
 			{server && server.urlString ? (
-				<NativeShellWebView
-					ref={webview}
-					style={webviewStyle}
-					containerStyle={webviewStyle}
-					refreshControlProps={{
-						// iOS colors
-						tintColor: theme.colors.grey1,
-						backgroundColor: theme.colors.grey0,
-						// Android colors
-						colors: [theme.colors.primary, theme.colors.secondary],
-						progressBackgroundColor: theme.colors.background
-					}}
-					// Error screen is displayed if loading fails
-					renderError={errorCode => (
-						<ErrorView
-							icon={{
-								name: 'cloud-off',
-								type: 'material'
-							}}
-							heading={t([`home.errors.${errorCode}.heading`, 'home.errors.offline.heading'])}
-							message={t([`home.errors.${errorCode}.description`, 'home.errors.offline.description'])}
-							details={[
-								t('home.errorCode', { errorCode }),
-								t('home.errorUrl', { url: server.urlString })
-							]}
-							buttonIcon={{
-								name: getIconName('refresh'),
-								type: 'ionicon'
-							}}
-							buttonTitle={t('home.retry')}
-							onPress={() => webview.current?.reload()}
-						/>
-					)}
-					// Loading screen is displayed when refreshing
-					renderLoading={() => <View style={styles.container} />}
-					// Update state on loading error
-					onError={({ nativeEvent: state }) => {
-						console.warn('Error', state);
-					}}
-					onHttpError={({ nativeEvent: state }) => {
-						console.warn('HTTP Error', state);
-						setHttpErrorStatus(state);
-					}}
-					onLoadStart={() => {
-						setIsLoading(true);
-						setHttpErrorStatus(null);
-					}}
-					// Update state when loading is complete
-					onLoadEnd={() => {
-						setIsLoading(false);
-					}}
-				/>
+				<>
+					<NativeShellWebView
+						ref={webview}
+						style={webviewStyle}
+						containerStyle={webviewStyle}
+						refreshControlProps={{
+							// iOS colors
+							tintColor: theme.colors.grey1,
+							backgroundColor: theme.colors.grey0,
+							// Android colors
+							colors: [theme.colors.primary, theme.colors.secondary],
+							progressBackgroundColor: theme.colors.background
+						}}
+						// Error screen is displayed if loading fails
+						renderError={errorCode => (
+							<ErrorView
+								icon={{
+									name: 'cloud-off',
+									type: 'material'
+								}}
+								heading={t([`home.errors.${errorCode}.heading`, 'home.errors.offline.heading'])}
+								message={t([`home.errors.${errorCode}.description`, 'home.errors.offline.description'])}
+								details={[
+									t('home.errorCode', { errorCode }),
+									t('home.errorUrl', { url: server.urlString })
+								]}
+								buttonIcon={{
+									name: getIconName('refresh'),
+									type: 'ionicon'
+								}}
+								buttonTitle={t('home.retry')}
+								onPress={() => webview.current?.reload()}
+							/>
+						)}
+						// Loading screen is displayed when refreshing
+						renderLoading={() => <View style={styles.container} />}
+						// Update state on loading error
+						onError={({ nativeEvent: state }) => {
+							console.warn('Error', state);
+						}}
+						onHttpError={({ nativeEvent: state }) => {
+							console.warn('HTTP Error', state);
+							setHttpErrorStatus(state);
+						}}
+						onLoadStart={() => {
+							setIsLoading(true);
+							setHttpErrorStatus(null);
+						}}
+						// Update state when loading is complete
+						onLoadEnd={() => {
+							setIsLoading(false);
+						}}
+					/>
+					<VideoPlayer/>
+				</>
 			) : (
 				<ErrorView
 					heading={t('home.errors.invalidServer.heading')}
